all mighty push! did auto code cleanup and fixed the bug where you cant go to host scene... next is Join as a guest and then link the game manager with the DMServer

+++ src/main/java/com/Seals/scrabble/model/hostSide/game/Tile.java
@@ -33,7 +33,7 @@ public class Tile {
     }
 
     int getPoints(char L) {
-        int index = (int) (L - 'A');
+        int index = L - 'A';
         return Values[index];
     }
